# Speed up schedule()