### Summary
Speed up ALAP and ASAP scheduling methods.

### Details and comments
This PR improves the implementation of scheduling algorithms (ALAP and ASAP) without changing their interface. It constructs the resulting Schedule at once in order to avoid useless creation of intermediate Schedules, which made the original schedule() slow.