@taalexander 2020-02-14
This makes sense to me that this would result in a speedup. I imagine the improvements could be very tangible. Do you have some temporary benchmarks we could so that we might consider merging this in before the full benchmark suite is ready?