@itoko 2020-02-14
Here is the improvement of time taken for schedule() in my local environment by this PR for a RB circuit with `n` Clifford gates. 
```
n=300: 22.656 seconds -> 0.937 seconds
n=200: 9.015 seconds -> 0.567 seconds
n=100: 2.076 seconds -> 0.282 seconds
```
We can see more improvements for deeper circuits as expected.

Benchmark code:
```
import numpy as np
from qiskit import transpile, schedule
from qiskit.ignis.verification import randomized_benchmarking as rb
from qiskit.test.mock.fake_openpulse_2q import FakeOpenPulse2Q

np.random.seed(10)
backend = FakeOpenPulse2Q()
rb_opts = {
    'length_vector': [n],
    'nseeds': 1,
    'rb_pattern': [[0, 1]]
}
rb_circs, xdata = rb.randomized_benchmarking_seq(**rb_opts)

qc_transpile = transpile(rb_circs[0], backend)

# measure the time taken only for schedule()
sched = schedule(qc_transpile, backend)
```
(Note: Please regard those circuits as random circuits for benchmark rather than RB circuits.)