@itoko 2020-02-14
One more thing: This fix #3678 as a by-product.